move UI components to separate subpackage
