+++ feature/circles/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/feature/circles/components/CircleItem.kt
@@ -1,4 +1,4 @@
-package com.livefast.eattrash.raccoonforfriendica.feature.circles.list
+package com.livefast.eattrash.raccoonforfriendica.feature.circles.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
